pyproject.toml
[project]
name = "knowledge-base-organizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = []
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.9",
    "typer>=0.19.2",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "hypothesis>=6.140.2",
    "isort>=6.0.1",
    "licensecheck>=2025.1.0",
    "lizard>=1.18.0",
    "mypy>=1.18.2",
    "nox==2025.02.09",
    "pip-audit>=2.9.0",
    "pre-commit>=4.3.0",
    "pre-commit-hooks>=6.0.0",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.2",
    "safety>=3.6.2",
    "types-pyyaml>=6.0.12.20250915",
]

doc = [
    "furo>=2025.9.25",
    "jupytext>=1.17.3",
    "myst-parser>=4.0.1",
    "nbsphinx>=0.9.7",
    "sphinx>=8.1.3",
    "sphinx-copybutton>=0.5.2",
]

[project.scripts]
knowledge-base-organizer = "knowledge_base_organizer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
